src/test/kotlin/com/clouddjr/advent2022/Day22Test.kt
package com.clouddjr.advent2022

import com.clouddjr.advent2022.Resources.resourceAsListOfString
import org.junit.jupiter.api.DisplayName
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Test
import kotlin.test.assertEquals

@DisplayName("Day 22")
class Day22Test {
    @Nested
    @DisplayName("Part 1")
    inner class Part1 {
        @Test
        fun `Matches actual`() {
            val answer = Day22(resourceAsListOfString("day22.txt")).solvePart1()

            assertEquals(97_356, answer)
        }
    }

    @Nested
    @DisplayName("Part 2")
    inner class Part2 {
        @Test
        fun `Matches actual`() {
            val answer = Day22(resourceAsListOfString("day22.txt")).solvePart2()

            assertEquals(120_175, answer)
        }
    }
}